add: insertMany, deprecates insertList & insertArray
